Created reusable NavAndHeader component and applied it to UserDashboard and TeamDashboard.
