+++ src/pages/UserDashboard.js
@@ -1,18 +1,30 @@
-import React from 'react'
-import { Layout } from 'antd';
-import DashboardNav from '../components/DashboardNav';
-import UserDashboardContent from '../components/UserDashboardContent';
+import React from 'react';
+import { connect } from "react-redux";
+
+import NavAndHeader from "../components/utils/NavAndHeader";
+import TeamList from '../components/UserTeamsList';
+import UserVideos from "../components/UserVideosList";
+
+import { Card } from 'antd';
 
 function UserDashboard() {
 
 	return (
-		<>
-			<Layout>
-				<DashboardNav />
-				<UserDashboardContent />
-			</Layout>
-		</>
+		<NavAndHeader>
+			<Card title="Your Teams" style={{ margin: "20px" }}>
+				<TeamList />
+			</Card>
+			<Card title="Your Videos" style={{ margin: "20px" }}>
+				<UserVideos />
+			</Card>
+		</NavAndHeader>
 	)
 }
 
-export default UserDashboard
+const mapStateToProps = (state) => {
+	return {
+		username: state.User.username,
+	}
+}
+
+export default connect(mapStateToProps, {})(UserDashboard);
